Drag drop toggle now works for only files

+++ src/main/java/com/i_comit/windows/Memory.java
@@ -99,10 +99,11 @@ public class Memory {
     }
 
     public static void saveIPAddress() {
-        if (!Main.adminBool) {
-            File miscsFile = new File(root + Main.masterFolder + "app\\.miscs.txt");
-            miscsFile.setReadable(true);
-            miscsFile.setWritable(true);
+        File miscsFile = new File(root + Main.masterFolder + "app\\.miscs.txt");
+        miscsFile.setReadable(true);
+        miscsFile.setWritable(true);
+        if (!Main.adminBool && miscsFile.exists()) {
+
             try {
                 FileWriter myWriter = new FileWriter(root + Main.masterFolder + "app\\.miscs.txt");
                 myWriter.write(Miscs.stringToHex(Main.jClientIPInput.getText()));
